requirements.txt
# Production Dependencies
# ======================

# Database connectivity
psycopg2-binary>=2.9.0
pyodbc>=4.0.32
asyncpg>=0.27.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0
imbalanced-learn>=0.9.0
pandas>=1.5.0
numpy>=1.21.0

# Rule Processing
pyDatalog>=0.17.1

# Configuration and Utilities
PyYAML>=6.0
cryptography>=3.4.8

# Monitoring and Metrics
prometheus-client>=0.15.0
psutil>=5.9.0

# Web Framework (for API endpoints)
fastapi>=0.88.0
uvicorn>=0.20.0

# Async Processing
aiohttp>=3.8.0
aiofiles>=22.1.0

# Caching (optional)
redis>=4.3.0
pymemcache>=4.0.0

# Data Processing
mlxtend>=0.21.0
tqdm>=4.64.0

# Serialization
joblib>=1.2.0

# Development Dependencies
# =======================

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0

# Code Quality
black>=22.0.0
flake8>=6.0.0
mypy>=0.991
isort>=5.11.0

# Documentation
sphinx>=5.3.0
sphinx-rtd-theme>=1.2.0

# Performance Testing
locust>=2.14.0

# Security
bandit>=1.7.0
safety>=2.3.0

# Additional Production Dependencies
# =================================

# Logging
structlog>=22.3.0

# Email
smtplib-ssl>=1.0.0

# HTTP Client
requests>=2.28.0

# Environment Management
python-dotenv>=0.21.0

# Data Validation
pydantic>=1.10.0

# Task Queue (alternative to async)
celery>=5.2.0

# Message Queue
pika>=1.3.0  # RabbitMQ
confluent-kafka>=1.9.0  # Kafka

# Streaming
apache-beam>=2.43.0

# Cloud Storage (if needed)
boto3>=1.26.0  # AWS
azure-storage-blob>=12.14.0  # Azure
google-cloud-storage>=2.7.0  # GCP

# API Documentation
swagger-ui-bundle>=0.0.9

# Health Checks
healthcheck>=1.3.3

# Rate Limiting
slowapi>=0.1.9

# JWT Authentication (if needed)
PyJWT>=2.6.0

# Container Support
gunicorn>=20.1.0

# Development Tools
# ================

# Jupyter (for analysis)
jupyter>=1.0.0
jupyterlab>=3.5.0

# Profiling
py-spy>=0.3.14
memory-profiler>=0.60.0

# Database Migration
alembic>=1.9.0

# Type Checking
types-PyYAML>=6.0.0
types-requests>=2.28.0
types-redis>=4.3.0

# Linting
pylint>=2.15.0
pre-commit>=2.21.0

# Package Management
pip-tools>=6.12.0

# Version Information
# ===================
# Python: >=3.8,<3.12
# This system is tested with Python 3.8, 3.9, 3.10, and 3.11

# Operating System Support
# ========================
# - Windows 10/11 (primary target)
# - Linux (Ubuntu 20.04+, RHEL 8+)
# - macOS 11+ (development only)

# Database Requirements
# ====================
# PostgreSQL: >=12.0 (for claims data)
# SQL Server: >=2017 (for results storage)
# Redis: >=6.0 (optional, for caching)

# Hardware Recommendations
# ========================
# Minimum:
# - 8 GB RAM
# - 4 CPU cores
# - 100 GB storage
#
# Recommended:
# - 32 GB RAM
# - 16 CPU cores
# - 500 GB SSD storage
#
# High-performance:
# - 64 GB RAM
# - 32 CPU cores
# - 1 TB NVMe SSD